Moving to and from functions into implementations of trait From.
- Implemented for LilyPondNote
- Implemented for Note
