+++ src/lilypond_objects/lilypond_note.rs
@@ -1,8 +1,10 @@
 //! Abstractions for LilyPond-syntax `String` representations of notes.
 
-use crate::notation::note::Note;
-use crate::notation::pitch::{Accidental, NoteName, Octave};
-use crate::notation::rhythm::{DurationType, Length};
+use crate::notation::{
+    note::Note,
+    pitch::{Accidental, NoteName, Octave},
+    rhythm::{DurationType, Length},
+};
 use lazy_static::lazy_static;
 use regex::Regex;
 
@@ -117,22 +119,6 @@ impl LilyPondNote {
         &self.note
     }
 
-    /// Translate a note object into a lilypond note string
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use lilypond::lilypond_objects::lilypond_note::LilyPondNote;
-    /// use lilypond::notation::note::Note;
-    /// use lilypond::notation::pitch::NoteName;
-    ///
-    /// let ly_note = Note::new(NoteName::A).to_lilypond_note();
-    /// assert_eq!(ly_note.get_note(), "a4");
-    /// ```
-    pub fn from_note(note: Note) -> LilyPondNote {
-        note.to_lilypond_note()
-    }
-
     /// Get a capture from current note according to label `&str capture` in `LILYPOND_NOTE_REGEX`.
     ///
     /// Available capture group labels follow. See
@@ -161,136 +147,77 @@ impl LilyPondNote {
             .replace(&self.note, format!("${}", capture))
             .to_string()
     }
+}
 
-    fn get_duration_type(&self) -> DurationType {
-        match self.get_capture("note_name").as_str() {
-            "r" => DurationType::Rest,
-            _ => DurationType::Note,
-        }
-    }
-
-    fn get_note_name(&self) -> NoteName {
-        match self.get_duration_type() {
-            DurationType::Rest => NoteName::None,
-            DurationType::Note => match self.get_capture("note_name").as_str() {
-                "a" => NoteName::A,
-                "b" => NoteName::B,
-                "c" => NoteName::C,
-                "d" => NoteName::D,
-                "e" => NoteName::E,
-                "f" => NoteName::F,
-                "g" => NoteName::G,
-                e => panic!("Invalid note name '{}'.", e),
-            },
-        }
-    }
-
-    fn get_accidental(&self) -> Accidental {
-        match self.get_duration_type() {
-            DurationType::Rest => Accidental::None,
-            DurationType::Note => match self.get_capture("accidental").as_str() {
-                "" => Accidental::None,
-                "s" => Accidental::Sharp,
-                "is" => Accidental::Sharp,
-                "ss" => Accidental::DoubleSharp,
-                "isis" => Accidental::DoubleSharp,
-                "f" => Accidental::Flat,
-                "es" => Accidental::Flat,
-                "ff" => Accidental::DoubleFlat,
-                "eses" => Accidental::DoubleFlat,
-                e => panic!("Invalid accidental '{}'.", e),
-            },
-        }
-    }
-
-    fn get_octave(&self) -> Octave {
-        match self.get_duration_type() {
-            DurationType::Rest => Octave::None,
-            DurationType::Note => {
-                // octave has to be usize to add count() results from it
-                let mut octave_int: usize = 3;
-                let octave_string = self.get_capture("octave");
-
-                if octave_string.contains(",") && octave_string.contains("'") {
-                    // Check for both octave transposition characters and panic
-                    panic!("Mixed octave transpostion symbols , and '.");
-                } else if octave_string.contains("'") {
-                    octave_int += octave_string.matches("'").count();
-                } else if octave_string.contains(",") {
-                    octave_int -= octave_string.matches(",").count();
-                }
-
-                match octave_int {
-                    0 => Octave::S0,
-                    1 => Octave::S1,
-                    2 => Octave::S2,
-                    3 => Octave::S3,
-                    4 => Octave::S4,
-                    5 => Octave::S5,
-                    6 => Octave::S6,
-                    7 => Octave::S7,
-                    8 => Octave::S8,
-                    9 => Octave::S9,
-                    _ => panic!("Invalid number of octave transpositions."),
-                }
-            }
-        }
-    }
-
-    fn get_length(&self) -> Length {
-        match self.get_capture("duration").as_str() {
-            "1" => Length::Whole,
-            "2" => Length::Half,
-            "4" => Length::Quarter,
-            "8" => Length::Eighth,
-            "16" => Length::Sixteenth,
-            "32" => Length::ThirtySecond,
-            "64" => Length::SixtyFourth,
-            "128" => Length::OneTwentyEighth,
-            "" => Default::default(),
-            e => panic!("Invalid duration '{}'.", e),
-        }
-    }
-
-    fn get_dot(&self) -> bool {
-        self.get_capture("dot") == "."
-    }
-
-    /// Translate a lilypond note string to a note object.
+impl From<&Note> for LilyPondNote {
+    /// Translate a note object into a lilypond note string
     ///
     /// # Examples
     ///
     /// ```rust
     /// use lilypond::lilypond_objects::lilypond_note::LilyPondNote;
     /// use lilypond::notation::note::Note;
-    /// use lilypond::notation::pitch::{Pitch, NoteName, Accidental, Octave};
-    /// use lilypond::notation::rhythm::{Rhythm, Length, DurationType};
+    /// use lilypond::notation::pitch::NoteName;
     ///
-    /// let note = LilyPondNote::new("af,8.").unwrap().to_note();
-    ///
-    /// assert_eq!(note.pitch.note_name, NoteName::A);
-    /// assert_eq!(note.pitch.octave, Octave::S2);
-    /// assert_eq!(note.pitch.accidental, Accidental::Flat);
-    /// assert_eq!(note.rhythm.length, Length::Eighth);
-    /// assert_eq!(note.rhythm.dotted, true);
-    /// assert_eq!(note.rhythm.duration_type, DurationType::Note);
+    /// let note = Note::new(NoteName::A);
+    /// let ly_note = LilyPondNote::from(&note);
+    /// assert_eq!(ly_note.get_note(), "a4");
     /// ```
-    pub fn to_note(&self) -> Note {
-        let note_name = self.get_note_name();
-        let note_accidental = self.get_accidental();
-        let note_octave = self.get_octave();
-        let note_length = self.get_length();
-        let note_duration_type = self.get_duration_type();
-        let note_dot = self.get_dot();
-
-        let mut note = Note::new(note_name);
-        note.pitch.accidental(note_accidental);
-        note.pitch.octave(note_octave);
-        note.rhythm.duration_type(note_duration_type);
-        note.rhythm.length(note_length);
-        note.rhythm.dotted(note_dot);
-
-        return note;
+    fn from(note: &Note) -> Self {
+        LilyPondNote::new(
+            format!(
+                "{}{}{}{}{}",
+                match note.rhythm.duration_type {
+                    DurationType::Rest => "r",
+                    DurationType::Note => match note.pitch.note_name {
+                        NoteName::A => "a",
+                        NoteName::B => "b",
+                        NoteName::C => "c",
+                        NoteName::D => "d",
+                        NoteName::E => "e",
+                        NoteName::F => "f",
+                        NoteName::G => "g",
+                        NoteName::None => "r",
+                    },
+                },
+                match note.pitch.accidental {
+                    Accidental::None => "",
+                    Accidental::Flat => "f",
+                    Accidental::DoubleFlat => "ff",
+                    Accidental::Sharp => "s",
+                    Accidental::DoubleSharp => "ss",
+                },
+                match note.pitch.octave {
+                    Octave::S0 => ",,,",
+                    Octave::S1 => ",,",
+                    Octave::S2 => ",",
+                    Octave::S3 => "",
+                    Octave::S4 => "'",
+                    Octave::S5 => "''",
+                    Octave::S6 => "'''",
+                    Octave::S7 => "''''",
+                    Octave::S8 => "'''''",
+                    Octave::S9 => "''''''",
+                    Octave::None => "",
+                },
+                match note.rhythm.length {
+                    Length::Whole => "1",
+                    Length::Half => "2",
+                    Length::Quarter => "4",
+                    Length::Eighth => "8",
+                    Length::Sixteenth => "16",
+                    Length::ThirtySecond => "32",
+                    Length::SixtyFourth => "64",
+                    Length::OneTwentyEighth => "128",
+                },
+                match note.rhythm.dotted {
+                    true => ".",
+                    false => "",
+                }
+            )
+            .as_str(),
+        )
+        .unwrap()
     }
 }
 
@@ -376,72 +303,8 @@ mod test {
     #[test]
     fn test_from_note() {
         let note = Note::new(NoteName::A);
-        let ly_note = LilyPondNote::from_note(note).note;
+        let ly_note = LilyPondNote::from(&note).note;
         assert_eq!("a4", ly_note);
-    }
-    #[test]
-    fn test_get_duration_type() {
-        let ly_note = LilyPondNote::new("r8").unwrap();
-        let duration_type = ly_note.get_duration_type();
-        assert_eq!(duration_type, DurationType::Rest);
-        let ly_note = LilyPondNote::new("f8").unwrap();
-        let duration_type = ly_note.get_duration_type();
-        assert_eq!(duration_type, DurationType::Note);
-    }
-    #[test]
-    fn test_get_note_name() {
-        let ly_note = LilyPondNote::new("r8").unwrap();
-        let duration_type = ly_note.get_note_name();
-        assert_eq!(duration_type, NoteName::None);
-        let ly_note = LilyPondNote::new("f8").unwrap();
-        let duration_type = ly_note.get_note_name();
-        assert_eq!(duration_type, NoteName::F);
-    }
-    #[test]
-    fn test_get_accidental() {
-        let ly_note = LilyPondNote::new("r8").unwrap();
-        let accidental_type = ly_note.get_accidental();
-        assert_eq!(accidental_type, Accidental::None);
-        let ly_note = LilyPondNote::new("fs").unwrap();
-        let accidental_type = ly_note.get_accidental();
-        assert_eq!(accidental_type, Accidental::Sharp);
-        let ly_note = LilyPondNote::new("ef").unwrap();
-        let accidental_type = ly_note.get_accidental();
-        assert_eq!(accidental_type, Accidental::Flat);
-    }
-    #[test]
-    fn test_get_octave() {
-        let ly_note = LilyPondNote::new("r8").unwrap();
-        let octave = ly_note.get_octave();
-        assert_eq!(octave, Octave::None);
-        let ly_note = LilyPondNote::new("fs,,,").unwrap();
-        let octave = ly_note.get_octave();
-        assert_eq!(octave, Octave::S0);
-        let ly_note = LilyPondNote::new("ef").unwrap();
-        let octave = ly_note.get_octave();
-        assert_eq!(octave, Octave::S3);
-        let ly_note = LilyPondNote::new("d''''''").unwrap();
-        let octave = ly_note.get_octave();
-        assert_eq!(octave, Octave::S9);
-    }
-    #[test]
-    fn test_get_length() {
-        let ly_note = LilyPondNote::new("r8").unwrap();
-        let length = ly_note.get_length();
-        assert_eq!(length, Length::Eighth);
-        let ly_note = LilyPondNote::new("as,128").unwrap();
-        let length = ly_note.get_length();
-        assert_eq!(length, Length::OneTwentyEighth);
-        let ly_note = LilyPondNote::new("bf''''64").unwrap();
-        let length = ly_note.get_length();
-        assert_eq!(length, Length::SixtyFourth);
-    }
-    #[test]
-    fn get_dot() {
-        let ly_note = LilyPondNote::new("r8.").unwrap();
-        assert!(ly_note.get_dot());
-        let ly_note = LilyPondNote::new("r8").unwrap();
-        assert!(!ly_note.get_dot());
     }
     #[test]
     fn test_to_note() {
@@ -458,7 +321,7 @@ mod test {
                 dotted: true,
             },
         };
-        assert_eq!(ly_note.to_note(), test_note);
+        assert_eq!(<&LilyPondNote as Into<Note>>::into(&ly_note), test_note);
         let ly_note = LilyPondNote::new("ef,,,64").unwrap();
         let test_note = Note {
             pitch: Pitch {
@@ -472,6 +335,6 @@ mod test {
                 dotted: false,
             },
         };
-        assert_eq!(ly_note.to_note(), test_note);
+        assert_eq!(<&LilyPondNote as Into<Note>>::into(&ly_note), test_note);
     }
 }
